server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
pydantic==2.3.0
python-dotenv==1.0.0
redis==4.6.0
rq==1.15.1
tenacity==8.2.3